prevent flickering while forcing max draw and extend to entire track

+++ PROJECTS/ROLLER/loadtrak.h
@@ -8,10 +8,10 @@
 typedef struct
 {
   int16 nForwardExtraStart;
-  uint8 byForwardMainChunks;
+  int16 byForwardMainChunks;
   uint8 byForwardExtraChunks;
   int16 nBackwardExtraStart;
-  uint8 byBackwardMainChunks;
+  int16 byBackwardMainChunks;
   uint8 byBackwardExtraChunks;
 } tTrakView;
